feat(cli): specify multiple stacks for deploy
Destroy is not yet supported but can be passed multiple stacks
