+++ packages/cdktf-cli/bin/cmds/destroy.ts
@@ -6,13 +6,13 @@ import { Errors } from "../../lib/errors";
 const config = cfg.readConfigSync();
 
 class Command implements yargs.CommandModule {
-  public readonly command = "destroy [stack] [OPTIONS]";
-  public readonly describe = "Destroy the given stack";
+  public readonly command = "destroy [OPTIONS] <stacks..>";
+  public readonly describe = "Destroy the given stacks";
 
   public readonly builder = (args: yargs.Argv) =>
     args
-      .positional("stack", {
-        desc: "Destroy stack which matches the given id only. Required when more than one stack is present in the app",
+      .positional("stacks", {
+        desc: "Destroy stacks matching the given ids. Required when more than one stack is present in the app",
         type: "string",
       })
       .option("app", {
