+++ packages/cdktf-cli/bin/cmds/ui/destroy.tsx
@@ -6,20 +6,20 @@ import { StreamView, StatusBottomBar, ApproveBottomBar } from "./components";
 
 interface DestroyConfig {
   outDir: string;
-  targetStack?: string;
+  targetStacks?: string[];
   synthCommand: string;
   autoApprove: boolean;
 }
 
 export const Destroy = ({
   outDir,
-  targetStack,
+  targetStacks,
   synthCommand,
   autoApprove,
 }: DestroyConfig): React.ReactElement => {
   const { projectUpdate, logEntries, done } = useCdktfProject(
     { outDir, synthCommand },
-    (project) => project.destroy({ stackName: targetStack, autoApprove })
+    (project) => project.destroy({ stackNames: targetStacks, autoApprove })
   );
 
   const bottomBar =
