+++ packages/cdktf-cli/lib/cdktf-stack.ts
@@ -86,7 +86,7 @@ export class CdktfStack {
   public stateMachine: InterpreterFrom<typeof stackExecutionMachine>;
   public stackName: string;
   public currentPlan?: TerraformPlan;
-  public stacks?: SynthesizedStack[];
+  public stack: SynthesizedStack;
   public outputs?: Record<string, any>;
   public outputsByConstructId?: NestedTerraformOutputs;
   public deployingResources?: DeployingResource[];
@@ -104,6 +104,7 @@ export class CdktfStack {
     autoApprove?: boolean;
     abortSignal: AbortSignal;
   }) {
+    this.stack = stack;
     this.stackName = stack.name;
     this.stateMachine = interpret(
       stackExecutionMachine.withContext({
