+++ packages/cdktf-cli/test/lib/fixtures/main.ts.fixture
@@ -34,6 +34,8 @@ export class HelloTerra extends TerraformStack {
 }
 
 const app = Testing.stubVersion(new App({ stackTraces: false }));
-new HelloTerra(app, "first");
+const first = new HelloTerra(app, "first");
 new HelloTerra(app, "second");
+const third = new HelloTerra(app, "third");
+third.addDependency(first);
 app.synth();
