+++ test/typescript/asset/test.ts
@@ -121,7 +121,7 @@ describe("full integration test", () => {
     expect(await driver.diff("normal")).toContain(
       "1 to add, 0 to change, 0 to destroy"
     );
-    await driver.deploy("normal");
+    await driver.deploy(["normal"]);
     expect(await driver.diff("normal")).toContain("No changes");
   });
 });
